Update NN-Vantage-Alert-Workflow.xml
Removed original line 26:
    <Initialize path="/first_run_alert" value="true" />

Removed future reference of the first run parameter.  

The PostEvents action was moved into a new IF condition that basically runs if the count in /data is NOT 0.  It then increments the page by 1.  The pagination was not working correctly with a customer using the current version of the workflow.  This was modified directly by an engineer at Nozomi support.