Update NN-Vantage-Asset-Workflow.xml
Same changes as I did the Nozomi Alerts workflow, but for the Assets endpoint.  

Essentially replaced all "alert" with "asset" in the workflow.  Removed the line about "first_run" value as it is no longer used. 

Put the PostEvents action in its own IF statement to run if the count in the number of assets is NOT 0, then increment the page count by 1.  The current pagination was not working in the workflow.  This change was made by a Nozomi engineer themselves.